Custom streams for HTTPUpdate
Background:
The user might want to manipulate or add headers to the OTA binaries.
Allowing the user to parse the TCPStream through a custom stream allows just that.
In my case I needed to be able to encrypt the binary for security reasons.

To do this I add a header with the IV of the AES encryption to the binary.
I then return that overhead of that header through the callback
with the new stream wrapper that handles the header parsing and decryption.

The overhead is then subtracted from the total binary length as it should not be written to flash.